moved a lot of files around and simplified. Setting up to put sign in form
