+++ public/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
 
 <head>
-  <meta name="viewport" content="width=device-width, user-scalable=no">
+  <meta name="viewport" content="width=device-width">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <meta charset="utf-8" />
   <link rel="shortcut icon" href="%PUBLIC_URL%/chef-icon.png" />
