+++ src/components/header/filters.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './styles/index.css';
+import '../styles/index.css';
 
 const Filters = ({ setMealTypeFilter, setDietaryNeedsFilter, mealTypeFilter, dietaryNeedsFilter, isBorked }) => {
   const mealTypeStrings = {
