+++ src/components/header/pie-box.js
@@ -1,6 +1,5 @@
 import React from 'react';
-import './styles/index.css';
-
+import '../styles/index.css';
 const PieBox = ({ setShowPie, showPie }) => {
   return (
       <div className="checkbox piebox">
