+++ src/components/recipe/recipe-list.js
@@ -6,13 +6,13 @@ import './styles.css';
 
 const RecipeList = ({ recipes, isBorked }) => {
   if (isEmpty(recipes)) {
-    return (<h3>{isBorked ? "Ouoops! Nu receepes! Bork Bork Bork!" : "Oops! No recipes!"}</h3>)
+    return (<p>{isBorked ? "Ouoops! Nu receepes! Bork Bork Bork!" : "Oops! No recipes!"}</p>)
   }
 
   const recipesWithAlphadivider = alphadividerUtil({ recipes, withRecipe: true });
   const list = recipesWithAlphadivider.map((heading, index) => {
     if (heading.length === 1) {
-      return (<h3 className='container bg-dark-blue text-cream' id={heading} key={index}>{heading.toUpperCase()}</h3>)
+      return (<h2 className='container bg-dark-blue text-cream' id={heading} key={index}>{heading.toUpperCase()}</h2>)
     }
     return (<Recipe recipe={recipes[heading]} key={index} />)
   });
