--- src/components/sign-in/index.js
@@ -1,12 +1,0 @@
-import React from 'react';
-
-const SignIn = ({}) => {
-    return (
-      <form>
-
-      </form>
-    )
-}
-
-export default SignIn;
-
